+++ src/pull.rs
@@ -1,44 +1,173 @@
 use std::f32;
 
 use nalgebra;
-use nalgebra::{Point2, Point3, Rotation3, Vector3};
+use nalgebra::{Point3, Rotation3, Vector3};
 
 use crate::convert::cast_usize;
 use crate::geometry;
 use crate::mesh::{Face, Mesh, UnorientedEdge};
 use crate::plane::Plane;
 
-/// Compute barycentric coordinates of point P in triangle A, B, C. Returns None
-/// for degenerate triangles.
-pub fn compute_barycentric_coords(
-    a: Point2<f32>,
-    b: Point2<f32>,
-    c: Point2<f32>,
-    p: Point2<f32>,
-) -> Option<Point3<f32>> {
-    let ab = b - a;
-    let ac = c - a;
-    let pa = a - p;
-    let xs = Vector3::new(ac.x, ab.x, pa.x);
-    let ys = Vector3::new(ac.y, ab.y, pa.y);
-    let ortho = xs.cross(&ys);
-    if f32::abs(ortho.z) < 1.0 {
-        None
+#[derive(Debug, Clone, Copy, PartialEq)]
+pub struct PulledPointWithDistance {
+    pub point: Point3<f32>,
+    pub distance: f32,
+}
+
+#[derive(Debug, Clone, Copy, PartialEq)]
+pub struct PointOnLine {
+    pub clamped: Point3<f32>,
+    pub unclamped: Point3<f32>,
+}
+
+/// Pulls arbitrary point to an arbitrary line.
+///
+/// Returns both, a clamped point on the line and an unclamped points on an
+/// endless ray.
+///
+/// https://stackoverflow.com/questions/3120357/get-closest-point-to-a-line
+pub fn pull_point_to_line(
+    point: &Point3<f32>,
+    line_start: &Point3<f32>,
+    line_end: &Point3<f32>,
+) -> PointOnLine {
+    if is_point_on_line_clamped(point, line_start, line_end) {
+        return PointOnLine {
+            clamped: *point,
+            unclamped: *point,
+        };
+    }
+    let start_to_point = point - line_start;
+    let start_to_end = line_end - line_start;
+
+    let square_magnitude_start_to_end = nalgebra::distance_squared(line_start, line_end);
+    let start_to_point_dot_start_to_end = start_to_point.dot(&start_to_end);
+    let line_parameter = start_to_point_dot_start_to_end / square_magnitude_start_to_end;
+
+    let closest_point_on_line_unclamped =
+        Point3::from(line_start.coords.lerp(&line_end.coords, line_parameter));
+    let closest_point_on_line_clamped = if line_parameter < 0.0 {
+        *line_start
+    } else if line_parameter > 1.0 {
+        *line_end
     } else {
-        Some(Point3::new(
-            1.0 - (ortho.x + ortho.y) / ortho.z,
-            ortho.y / ortho.z,
-            ortho.x / ortho.z,
-        ))
+        closest_point_on_line_unclamped
+    };
+
+    PointOnLine {
+        clamped: closest_point_on_line_clamped,
+        unclamped: closest_point_on_line_unclamped,
     }
 }
 
-/// Checks if all three points of a triangle lay on the same line.
-fn are_vertices_colinear(v0: &Point3<f32>, v1: &Point3<f32>, v2: &Point3<f32>) -> bool {
-    let v0_normalized = v0.coords.normalize();
-    let v1_normalized = v1.coords.normalize();
-    let v2_normalized = v2.coords.normalize();
-    v0_normalized == v1_normalized && v0_normalized == v2_normalized
+/// Pulls arbitrary point to the closest point of a plane.
+///
+/// Cast a ray from the point perpendicular to the plane and calculate their
+/// intersection.
+#[allow(dead_code)]
+pub fn pull_point_to_plane(point: &Point3<f32>, plane: &Plane) -> PulledPointWithDistance {
+    if plane.contains_point(point) {
+        PulledPointWithDistance {
+            point: *point,
+            distance: 0.0,
+        }
+    } else {
+        ray_intersects_plane(point, &plane.normal(), plane)
+            .expect("The normal is parallel to its plane")
+    }
+}
+
+/// Pulls arbitrary point to the closest point of a mesh geometry.
+///
+/// Cast a ray from the point perpendicular to each mesh face and if there is an
+/// intersection, add it to the list. Also measure a distance to the closest
+/// point on each mesh geometry edge. Pick the closest point of them all as the
+/// pulled point.
+#[allow(dead_code)]
+pub fn pull_point_to_mesh(
+    point: &Point3<f32>,
+    mesh: &Mesh,
+    unoriented_edges: &[UnorientedEdge],
+) -> PulledPointWithDistance {
+    let vertices = mesh.vertices();
+    let all_mesh_faces_with_normals = mesh.faces().iter().map(|Face::Triangle(t_f)| {
+        let face_vertices = (
+            &vertices[cast_usize(t_f.vertices.0)],
+            &vertices[cast_usize(t_f.vertices.1)],
+            &vertices[cast_usize(t_f.vertices.2)],
+        );
+        (
+            face_vertices,
+            geometry::compute_triangle_normal(face_vertices.0, face_vertices.1, face_vertices.2),
+        )
+    });
+    let mut is_on_mesh = false;
+
+    let mut pulled_points: Vec<PulledPointWithDistance> = Vec::new();
+    // Pull to faces
+    for (face_vertices, face_normal) in all_mesh_faces_with_normals {
+        // If the point already lays in the face, it means it's already puled to the mesh.
+        if is_point_in_triangle(point, face_vertices.0, face_vertices.1, face_vertices.2) {
+            is_on_mesh = true;
+            break;
+        }
+        // If triangle vertices are colinear, it's enough to pull to
+        // triangle edges later on.
+        if !geometry::are_points_colinear(face_vertices.0, face_vertices.1, face_vertices.2) {
+            if let Some(intersection_point) = ray_intersects_triangle(
+                point,
+                &(-1.0 * face_normal),
+                face_vertices.0,
+                face_vertices.1,
+                face_vertices.2,
+            ) {
+                pulled_points.push(intersection_point);
+            }
+        }
+    }
+
+    // Exit and return the point itself.
+    if is_on_mesh {
+        return PulledPointWithDistance {
+            point: *point,
+            distance: 0.0,
+        };
+    }
+
+    // Pull to edges
+    for u_e in unoriented_edges {
+        let closest_point = pull_point_to_line(
+            point,
+            &vertices[cast_usize(u_e.0.vertices.0)],
+            &vertices[cast_usize(u_e.0.vertices.1)],
+        );
+        // If the point already lays in the edge, it means it's already puled to the mesh.
+        if closest_point.clamped == *point {
+            is_on_mesh = true;
+            break;
+        }
+        pulled_points.push(PulledPointWithDistance {
+            point: closest_point.clamped,
+            distance: nalgebra::distance(point, &closest_point.clamped),
+        });
+    }
+
+    // Exit and return the point itself.
+    if is_on_mesh {
+        return PulledPointWithDistance {
+            point: *point,
+            distance: 0.0,
+        };
+    }
+
+    let mut closest_pulled_point = pulled_points.pop().expect("No pulled point found");
+    for current_pulled_point in pulled_points {
+        if current_pulled_point.distance < closest_pulled_point.distance {
+            closest_pulled_point = current_pulled_point;
+        }
+    }
+
+    closest_pulled_point
 }
 
 /// Checks if a point lies in a triangle.
@@ -67,14 +196,14 @@ fn is_point_in_triangle(
 
     // If the triangle is degenerated into a line, check if the point lies on
     // the line.
-    if are_vertices_colinear(triangle_vertex0, triangle_vertex1, triangle_vertex2) {
+    if geometry::are_points_colinear(triangle_vertex0, triangle_vertex1, triangle_vertex2) {
         return is_point_on_line_clamped(point, triangle_vertex0, triangle_vertex1);
     }
 
     let plane = Plane::from_three_points(triangle_vertex0, triangle_vertex1, triangle_vertex2);
 
     // If the point is not on the triangle plane, it's also not on the triangle.
-    if !is_point_on_plane(point, &plane) {
+    if !plane.contains_point(&point) {
         return false;
     }
 
@@ -111,7 +240,7 @@ fn is_point_in_triangle(
             return false;
         };
 
-    let barycentric_point = compute_barycentric_coords(
+    let barycentric_point = geometry::compute_barycentric_coords(
         horizontal_vertex0.xy(),
         horizontal_vertex1.xy(),
         horizontal_vertex2.xy(),
@@ -129,12 +258,6 @@ fn is_point_in_triangle(
             barycentric_point.x + barycentric_point.y + barycentric_point.z,
             1.0
         )
-}
-
-#[derive(Debug, Clone, Copy, PartialEq)]
-pub struct PulledPointWithDistance {
-    point: Point3<f32>,
-    distance: f32,
 }
 
 /// The Möller–Trumbore ray-triangle intersection algorithm is a fast method for
@@ -197,20 +320,6 @@ fn ray_intersects_triangle(
     }
 }
 
-/// Test if an arbitrary point lies on a plane.
-///
-/// https://stackoverflow.com/questions/17227149/using-dot-product-to-determine-if-point-lies-on-a-plane
-fn is_point_on_plane(point: &Point3<f32>, plane: &Plane) -> bool {
-    let vector_from_plane_point_to_point = point - plane.origin();
-    *point == plane.origin()
-        || approx::relative_eq!(
-            vector_from_plane_point_to_point
-                .normalize()
-                .dot(&plane.normal()),
-            0.0
-        )
-}
-
 /// Find an intersection or a ray and a plane.
 ///
 /// #Panics
@@ -239,12 +348,6 @@ fn ray_intersects_plane(
             distance: t_parameter.abs(),
         })
     }
-}
-
-#[derive(Debug, Clone, Copy, PartialEq)]
-pub struct PointOnLine {
-    clamped: Point3<f32>,
-    unclamped: Point3<f32>,
 }
 
 /// Checks if a point lies on an infinite line.
@@ -280,156 +383,6 @@ fn is_point_on_line_clamped(
             && nalgebra::distance_squared(line_end, point) <= line_length_squared
 }
 
-/// Pulls arbitrary point to an arbitrary line.
-///
-/// Returns both, a clamped point on the line and an unclamped points on an
-/// endless ray.
-///
-/// https://stackoverflow.com/questions/3120357/get-closest-point-to-a-line
-pub fn pull_point_to_line(
-    point: &Point3<f32>,
-    line_start: &Point3<f32>,
-    line_end: &Point3<f32>,
-) -> PointOnLine {
-    if is_point_on_line_clamped(point, line_start, line_end) {
-        return PointOnLine {
-            clamped: *point,
-            unclamped: *point,
-        };
-    }
-    let start_to_point = point - line_start;
-    let start_to_end = line_end - line_start;
-
-    let square_magnitude_start_to_end = nalgebra::distance_squared(line_start, line_end);
-    let start_to_point_dot_start_to_end = start_to_point.dot(&start_to_end);
-    let line_parameter = start_to_point_dot_start_to_end / square_magnitude_start_to_end;
-
-    let closest_point_on_line_unclamped =
-        Point3::from(line_start.coords.lerp(&line_end.coords, line_parameter));
-    let closest_point_on_line_clamped = if line_parameter < 0.0 {
-        *line_start
-    } else if line_parameter > 1.0 {
-        *line_end
-    } else {
-        closest_point_on_line_unclamped
-    };
-
-    PointOnLine {
-        clamped: closest_point_on_line_clamped,
-        unclamped: closest_point_on_line_unclamped,
-    }
-}
-
-/// Pulls arbitrary point to the closest point of a mesh geometry.
-///
-/// Cast a ray from the point perpendicular to each mesh face and if there is an
-/// intersection, add it to the list. Also measure a distance to the closest
-/// point on each mesh geometry edge. Pick the closest point of them all as the
-/// pulled point.
-#[allow(dead_code)]
-pub fn pull_point_to_mesh(
-    point: &Point3<f32>,
-    mesh: &Mesh,
-    unoriented_edges: &[UnorientedEdge],
-) -> PulledPointWithDistance {
-    let vertices = mesh.vertices();
-    let all_mesh_faces_with_normals = mesh.faces().iter().map(|Face::Triangle(t_f)| {
-        let face_vertices = (
-            &vertices[cast_usize(t_f.vertices.0)],
-            &vertices[cast_usize(t_f.vertices.1)],
-            &vertices[cast_usize(t_f.vertices.2)],
-        );
-        (
-            face_vertices,
-            geometry::compute_triangle_normal(face_vertices.0, face_vertices.1, face_vertices.2),
-        )
-    });
-    let mut is_on_mesh = false;
-
-    let mut pulled_points: Vec<PulledPointWithDistance> = Vec::new();
-    // Pull to faces
-    for (face_vertices, face_normal) in all_mesh_faces_with_normals {
-        // If the point already lays in the face, it means it's already puled to the mesh.
-        if is_point_in_triangle(point, face_vertices.0, face_vertices.1, face_vertices.2) {
-            is_on_mesh = true;
-            break;
-        }
-        // If triangle vertices are colinear, it's enough to pull to
-        // triangle edges later on.
-        if !are_vertices_colinear(face_vertices.0, face_vertices.1, face_vertices.2) {
-            if let Some(intersection_point) = ray_intersects_triangle(
-                point,
-                &(-1.0 * face_normal),
-                face_vertices.0,
-                face_vertices.1,
-                face_vertices.2,
-            ) {
-                pulled_points.push(intersection_point);
-            }
-        }
-    }
-
-    // Exit and return the point itself.
-    if is_on_mesh {
-        return PulledPointWithDistance {
-            point: *point,
-            distance: 0.0,
-        };
-    }
-
-    // Pull to edges
-    for u_e in unoriented_edges {
-        let closest_point = pull_point_to_line(
-            point,
-            &vertices[cast_usize(u_e.0.vertices.0)],
-            &vertices[cast_usize(u_e.0.vertices.1)],
-        );
-        // If the point already lays in the edge, it means it's already puled to the mesh.
-        if closest_point.clamped == *point {
-            is_on_mesh = true;
-            break;
-        }
-        pulled_points.push(PulledPointWithDistance {
-            point: closest_point.clamped,
-            distance: nalgebra::distance(point, &closest_point.clamped),
-        });
-    }
-
-    // Exit and return the point itself.
-    if is_on_mesh {
-        return PulledPointWithDistance {
-            point: *point,
-            distance: 0.0,
-        };
-    }
-
-    let mut closest_pulled_point = pulled_points.pop().expect("No pulled point found");
-    for current_pulled_point in pulled_points {
-        if current_pulled_point.distance < closest_pulled_point.distance {
-            closest_pulled_point = current_pulled_point;
-        }
-    }
-
-    closest_pulled_point
-}
-
-/// Pulls arbitrary point to the closest point of a plane.
-///
-/// Cast a ray from the point perpendicular to the plane and calculate their
-/// intersection.
-#[allow(dead_code)]
-pub fn pull_point_to_plane(point: &Point3<f32>, plane: &Plane) -> PulledPointWithDistance {
-    if is_point_on_plane(point, &plane) {
-        PulledPointWithDistance {
-            point: *point,
-            distance: 0.0,
-        }
-    } else {
-        ray_intersects_plane(point, &plane.normal(), plane)
-            .expect("The normal is parallel to its plane")
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use crate::mesh::primitive;
@@ -437,55 +390,247 @@ mod tests {
     use super::*;
 
     #[test]
-    fn test_compute_barycentric_coords_for_point_inside() {
-        let triangle_points = (
-            Point2::new(0.0, 1.0),
-            Point2::new(-0.866025, -0.5),
-            Point2::new(0.866025, -0.5),
+    fn test_pull_point_to_line_for_point_pulled_to_center() {
+        let line_start = &Point3::new(-1.0, 0.0, 0.0);
+        let line_end = &Point3::new(1.0, 0.0, 0.0);
+        let test_point = Point3::new(0.0, 1.0, 1.0);
+
+        let point_on_line_clamped_correct = Point3::new(0.0, 0.0, 0.0);
+        let point_on_line_unclamped_correct = Point3::new(0.0, 0.0, 0.0);
+
+        let point_on_line_calculated = pull_point_to_line(&test_point, line_start, line_end);
+
+        assert_eq!(
+            point_on_line_clamped_correct,
+            point_on_line_calculated.clamped
         );
-
-        let test_point = Point2::new(0.0, 0.0);
-        let barycentric_calculated = compute_barycentric_coords(
-            triangle_points.0,
-            triangle_points.1,
-            triangle_points.2,
-            test_point,
-        )
-        .expect("Could not calculate the barycentric coords");
-
-        let barycentric_correct = Point3::new(0.333333, 0.333333, 0.333333);
-
-        assert!(barycentric_calculated.coords.relative_eq(
-            &barycentric_correct.coords,
-            0.001,
-            0.001
-        ));
+        assert_eq!(
+            point_on_line_unclamped_correct,
+            point_on_line_calculated.unclamped
+        );
     }
 
     #[test]
-    fn test_compute_barycentric_coords_for_point_outside() {
-        let triangle_points = (
-            Point2::new(0.0, 1.0),
-            Point2::new(-0.866025, -0.5),
-            Point2::new(0.866025, -0.5),
+    fn test_pull_point_to_line_for_point_pulled_below_start() {
+        let line_start = &Point3::new(-1.0, 0.0, 0.0);
+        let line_end = &Point3::new(1.0, 0.0, 0.0);
+        let test_point = Point3::new(-2.0, 1.0, 1.0);
+
+        let point_on_line_clamped_correct = Point3::new(-1.0, 0.0, 0.0);
+        let point_on_line_unclamped_correct = Point3::new(-2.0, 0.0, 0.0);
+
+        let point_on_line_calculated = pull_point_to_line(&test_point, line_start, line_end);
+
+        assert_eq!(
+            point_on_line_clamped_correct,
+            point_on_line_calculated.clamped
         );
+        assert_eq!(
+            point_on_line_unclamped_correct,
+            point_on_line_calculated.unclamped
+        );
+    }
 
-        let test_point = Point2::new(0.0, 2.0);
-        let barycentric_calculated = compute_barycentric_coords(
-            triangle_points.0,
-            triangle_points.1,
-            triangle_points.2,
-            test_point,
-        )
-        .expect("Could not calculate the barycentric coords");
+    #[test]
+    fn test_pull_point_to_line_for_point_pulled_beyond_end() {
+        let line_start = &Point3::new(-1.0, 0.0, 0.0);
+        let line_end = &Point3::new(1.0, 0.0, 0.0);
+        let test_point = Point3::new(2.0, 1.0, 1.0);
 
-        let barycentric_correct = Point3::new(1.6666667, -0.33333334, -0.33333334);
+        let point_on_line_clamped_correct = Point3::new(1.0, 0.0, 0.0);
+        let point_on_line_unclamped_correct = Point3::new(2.0, 0.0, 0.0);
 
-        assert!(barycentric_calculated.coords.relative_eq(
-            &barycentric_correct.coords,
+        let point_on_line_calculated = pull_point_to_line(&test_point, line_start, line_end);
+
+        assert_eq!(
+            point_on_line_clamped_correct,
+            point_on_line_calculated.clamped
+        );
+        assert_eq!(
+            point_on_line_unclamped_correct,
+            point_on_line_calculated.unclamped
+        );
+    }
+
+    #[test]
+    fn test_pull_point_to_line_for_point_on_the_line() {
+        let line_start = &Point3::new(-1.0, 0.0, 0.0);
+        let line_end = &Point3::new(1.0, 0.0, 0.0);
+        let test_point = Point3::new(0.0, 0.0, 0.0);
+
+        let point_on_line_clamped_correct = Point3::new(0.0, 0.0, 0.0);
+        let point_on_line_unclamped_correct = Point3::new(0.0, 0.0, 0.0);
+
+        let point_on_line_calculated = pull_point_to_line(&test_point, line_start, line_end);
+
+        assert_eq!(
+            point_on_line_clamped_correct,
+            point_on_line_calculated.clamped
+        );
+        assert_eq!(
+            point_on_line_unclamped_correct,
+            point_on_line_calculated.unclamped
+        );
+    }
+
+    #[test]
+    fn test_pull_point_to_plane_for_point_on_plane_and_horizontal_plane() {
+        let test_point: Point3<f32> = Point3::origin();
+        let test_plane =
+            Plane::from_origin_and_normal(&Point3::origin(), &Vector3::new(0.0, 0.0, 1.0));
+
+        let point_pulled_to_plane_calculated = pull_point_to_plane(&test_point, &test_plane);
+
+        assert_eq!(point_pulled_to_plane_calculated.point, test_point);
+        assert_eq!(point_pulled_to_plane_calculated.distance, 0.0);
+    }
+
+    #[test]
+    fn test_pull_point_to_plane_for_point_on_plane_and_vertical_plane() {
+        let test_point: Point3<f32> = Point3::origin();
+        let test_plane =
+            Plane::from_origin_and_normal(&Point3::origin(), &Vector3::new(1.0, 0.0, 0.0));
+
+        let point_pulled_to_plane_calculated = pull_point_to_plane(&test_point, &test_plane);
+
+        assert_eq!(point_pulled_to_plane_calculated.point, test_point);
+        assert_eq!(point_pulled_to_plane_calculated.distance, 0.0);
+    }
+
+    #[test]
+    fn test_pull_point_to_plane_for_point_above_plane_and_horizontal_plane() {
+        let test_point = Point3::new(0.0, 0.0, 1.0);
+        let test_plane =
+            Plane::from_origin_and_normal(&Point3::origin(), &Vector3::new(0.0, 0.0, 1.0));
+
+        let point_pulled_to_plane_calculated = pull_point_to_plane(&test_point, &test_plane);
+
+        assert_eq!(
+            point_pulled_to_plane_calculated.point,
+            Point3::new(0.0, 0.0, 0.0)
+        );
+        assert_eq!(point_pulled_to_plane_calculated.distance, 1.0);
+    }
+
+    #[test]
+    fn test_pull_point_to_plane_for_arbitrary_point_and_arbitrary_plane() {
+        let test_point = Point3::new(2.5, 4.2, 1.8);
+        let test_plane = Plane::from_origin_and_normal(&Point3::origin(), &test_point.coords);
+
+        let distance_correct = nalgebra::distance(&test_point, &Point3::origin());
+
+        let point_pulled_to_plane_calculated = pull_point_to_plane(&test_point, &test_plane);
+
+        assert!(point_pulled_to_plane_calculated.point.coords.relative_eq(
+            &Vector3::new(0.0, 0.0, 0.0),
             0.001,
-            0.001
+            0.001,
         ));
+        assert_eq!(point_pulled_to_plane_calculated.distance, distance_correct);
+    }
+
+    #[test]
+    fn test_pull_point_to_mesh_cube_point_inside_left() {
+        let cube = primitive::create_cube_sharp([0.0, 0.0, 0.0], 1.0);
+        let test_point = Point3::new(-0.25, 0.0, 0.0);
+        let unoriented_edges: Vec<_> = cube.unoriented_edges_iter().collect();
+
+        let point_on_mesh_correct = Point3::new(-1.0, 0.0, 0.0);
+        let pulled_point_on_mesh_calculated =
+            pull_point_to_mesh(&test_point, &cube, &unoriented_edges);
+
+        assert_eq!(point_on_mesh_correct, pulled_point_on_mesh_calculated.point);
+        assert_eq!(0.75, pulled_point_on_mesh_calculated.distance);
+    }
+
+    #[test]
+    fn test_pull_point_to_mesh_cube_point_inside_top_front_right() {
+        let cube = primitive::create_cube_sharp([0.0, 0.0, 0.0], 1.0);
+        let test_point = Point3::new(0.25, 0.25, 0.25);
+        let unoriented_edges: Vec<_> = cube.unoriented_edges_iter().collect();
+
+        // any of the following points on mesh would be correct
+        let points_on_mesh_correct = vec![
+            Point3::new(1.0, 0.25, 0.25),
+            Point3::new(0.25, 1.0, 0.25),
+            Point3::new(0.25, 0.25, 1.0),
+        ];
+        let pulled_point_on_mesh_calculated =
+            pull_point_to_mesh(&test_point, &cube, &unoriented_edges);
+
+        assert!(points_on_mesh_correct
+            .iter()
+            .any(|p| *p == pulled_point_on_mesh_calculated.point));
+
+        assert_eq!(0.75, pulled_point_on_mesh_calculated.distance);
+    }
+
+    #[test]
+    fn test_pull_point_to_mesh_cube_point_outside_top_front_right() {
+        let cube = primitive::create_cube_sharp([0.0, 0.0, 0.0], 1.0);
+        let test_point = Point3::new(1.25, 1.25, 1.25);
+        let unoriented_edges: Vec<_> = cube.unoriented_edges_iter().collect();
+
+        // corner
+        let point_on_mesh_correct = Point3::new(1.0, 1.0, 1.0);
+        let pulled_point_on_mesh_calculated =
+            pull_point_to_mesh(&test_point, &cube, &unoriented_edges);
+
+        assert_eq!(pulled_point_on_mesh_calculated.point, point_on_mesh_correct);
+        assert_eq!(
+            pulled_point_on_mesh_calculated.distance,
+            nalgebra::distance(&test_point, &pulled_point_on_mesh_calculated.point),
+        );
+    }
+
+    #[test]
+    fn test_pull_point_to_mesh_cube_point_outside_front_right() {
+        let cube = primitive::create_cube_sharp([0.0, 0.0, 0.0], 1.0);
+        let test_point = Point3::new(1.25, 1.25, 0.25);
+        let unoriented_edges: Vec<_> = cube.unoriented_edges_iter().collect();
+
+        // on the edge
+        let point_on_mesh_correct = Point3::new(1.0, 1.0, 0.25);
+        let pulled_point_on_mesh_calculated =
+            pull_point_to_mesh(&test_point, &cube, &unoriented_edges);
+
+        assert_eq!(pulled_point_on_mesh_calculated.point, point_on_mesh_correct);
+        assert_eq!(
+            pulled_point_on_mesh_calculated.distance,
+            nalgebra::distance(&test_point, &pulled_point_on_mesh_calculated.point),
+        );
+    }
+
+    #[test]
+    fn test_pull_point_to_mesh_cube_point_on_face() {
+        let cube = primitive::create_cube_sharp([0.0, 0.0, 0.0], 1.0);
+        let test_point = Point3::new(0.0, 0.0, 1.0);
+        let unoriented_edges: Vec<_> = cube.unoriented_edges_iter().collect();
+
+        let pulled_point_on_mesh_calculated =
+            pull_point_to_mesh(&test_point, &cube, &unoriented_edges);
+
+        assert_eq!(pulled_point_on_mesh_calculated.point, test_point);
+        assert_eq!(
+            pulled_point_on_mesh_calculated.distance,
+            nalgebra::distance(&test_point, &pulled_point_on_mesh_calculated.point),
+        );
+    }
+
+    #[test]
+    fn test_pull_point_to_mesh_cube_point_on_edge() {
+        let cube = primitive::create_cube_sharp([0.0, 0.0, 0.0], 1.0);
+        let test_point = Point3::new(1.0, 1.0, 0.0);
+        let unoriented_edges: Vec<_> = cube.unoriented_edges_iter().collect();
+
+        let pulled_point_on_mesh_calculated =
+            pull_point_to_mesh(&test_point, &cube, &unoriented_edges);
+
+        assert_eq!(pulled_point_on_mesh_calculated.point, test_point);
+        assert_eq!(
+            pulled_point_on_mesh_calculated.distance,
+            nalgebra::distance(&test_point, &pulled_point_on_mesh_calculated.point),
+        );
     }
 
     #[test]
@@ -832,214 +977,6 @@ mod tests {
     }
 
     #[test]
-    fn test_pull_point_to_line_for_point_pulled_to_center() {
-        let line_start = &Point3::new(-1.0, 0.0, 0.0);
-        let line_end = &Point3::new(1.0, 0.0, 0.0);
-        let test_point = Point3::new(0.0, 1.0, 1.0);
-
-        let point_on_line_clamped_correct = Point3::new(0.0, 0.0, 0.0);
-        let point_on_line_unclamped_correct = Point3::new(0.0, 0.0, 0.0);
-
-        let point_on_line_calculated = pull_point_to_line(&test_point, line_start, line_end);
-
-        assert_eq!(
-            point_on_line_clamped_correct,
-            point_on_line_calculated.clamped
-        );
-        assert_eq!(
-            point_on_line_unclamped_correct,
-            point_on_line_calculated.unclamped
-        );
-    }
-
-    #[test]
-    fn test_pull_point_to_line_for_point_pulled_below_start() {
-        let line_start = &Point3::new(-1.0, 0.0, 0.0);
-        let line_end = &Point3::new(1.0, 0.0, 0.0);
-        let test_point = Point3::new(-2.0, 1.0, 1.0);
-
-        let point_on_line_clamped_correct = Point3::new(-1.0, 0.0, 0.0);
-        let point_on_line_unclamped_correct = Point3::new(-2.0, 0.0, 0.0);
-
-        let point_on_line_calculated = pull_point_to_line(&test_point, line_start, line_end);
-
-        assert_eq!(
-            point_on_line_clamped_correct,
-            point_on_line_calculated.clamped
-        );
-        assert_eq!(
-            point_on_line_unclamped_correct,
-            point_on_line_calculated.unclamped
-        );
-    }
-
-    #[test]
-    fn test_pull_point_to_line_for_point_pulled_beyond_end() {
-        let line_start = &Point3::new(-1.0, 0.0, 0.0);
-        let line_end = &Point3::new(1.0, 0.0, 0.0);
-        let test_point = Point3::new(2.0, 1.0, 1.0);
-
-        let point_on_line_clamped_correct = Point3::new(1.0, 0.0, 0.0);
-        let point_on_line_unclamped_correct = Point3::new(2.0, 0.0, 0.0);
-
-        let point_on_line_calculated = pull_point_to_line(&test_point, line_start, line_end);
-
-        assert_eq!(
-            point_on_line_clamped_correct,
-            point_on_line_calculated.clamped
-        );
-        assert_eq!(
-            point_on_line_unclamped_correct,
-            point_on_line_calculated.unclamped
-        );
-    }
-
-    #[test]
-    fn test_pull_point_to_line_for_point_on_the_line() {
-        let line_start = &Point3::new(-1.0, 0.0, 0.0);
-        let line_end = &Point3::new(1.0, 0.0, 0.0);
-        let test_point = Point3::new(0.0, 0.0, 0.0);
-
-        let point_on_line_clamped_correct = Point3::new(0.0, 0.0, 0.0);
-        let point_on_line_unclamped_correct = Point3::new(0.0, 0.0, 0.0);
-
-        let point_on_line_calculated = pull_point_to_line(&test_point, line_start, line_end);
-
-        assert_eq!(
-            point_on_line_clamped_correct,
-            point_on_line_calculated.clamped
-        );
-        assert_eq!(
-            point_on_line_unclamped_correct,
-            point_on_line_calculated.unclamped
-        );
-    }
-
-    #[test]
-    fn test_pull_point_to_mesh_cube_point_inside_left() {
-        let cube = primitive::create_cube_sharp([0.0, 0.0, 0.0], 1.0);
-        let test_point = Point3::new(-0.25, 0.0, 0.0);
-        let unoriented_edges: Vec<_> = cube.unoriented_edges_iter().collect();
-
-        let point_on_mesh_correct = Point3::new(-1.0, 0.0, 0.0);
-        let pulled_point_on_mesh_calculated =
-            pull_point_to_mesh(&test_point, &cube, &unoriented_edges);
-
-        assert_eq!(point_on_mesh_correct, pulled_point_on_mesh_calculated.point);
-        assert_eq!(0.75, pulled_point_on_mesh_calculated.distance);
-    }
-
-    #[test]
-    fn test_pull_point_to_mesh_cube_point_inside_top_front_right() {
-        let cube = primitive::create_cube_sharp([0.0, 0.0, 0.0], 1.0);
-        let test_point = Point3::new(0.25, 0.25, 0.25);
-        let unoriented_edges: Vec<_> = cube.unoriented_edges_iter().collect();
-
-        // any of the following points on mesh would be correct
-        let points_on_mesh_correct = vec![
-            Point3::new(1.0, 0.25, 0.25),
-            Point3::new(0.25, 1.0, 0.25),
-            Point3::new(0.25, 0.25, 1.0),
-        ];
-        let pulled_point_on_mesh_calculated =
-            pull_point_to_mesh(&test_point, &cube, &unoriented_edges);
-
-        assert!(points_on_mesh_correct
-            .iter()
-            .any(|p| *p == pulled_point_on_mesh_calculated.point));
-
-        assert_eq!(0.75, pulled_point_on_mesh_calculated.distance);
-    }
-
-    #[test]
-    fn test_pull_point_to_mesh_cube_point_outside_top_front_right() {
-        let cube = primitive::create_cube_sharp([0.0, 0.0, 0.0], 1.0);
-        let test_point = Point3::new(1.25, 1.25, 1.25);
-        let unoriented_edges: Vec<_> = cube.unoriented_edges_iter().collect();
-
-        // corner
-        let point_on_mesh_correct = Point3::new(1.0, 1.0, 1.0);
-        let pulled_point_on_mesh_calculated =
-            pull_point_to_mesh(&test_point, &cube, &unoriented_edges);
-
-        assert_eq!(pulled_point_on_mesh_calculated.point, point_on_mesh_correct);
-        assert_eq!(
-            pulled_point_on_mesh_calculated.distance,
-            nalgebra::distance(&test_point, &pulled_point_on_mesh_calculated.point),
-        );
-    }
-
-    #[test]
-    fn test_pull_point_to_mesh_cube_point_outside_front_right() {
-        let cube = primitive::create_cube_sharp([0.0, 0.0, 0.0], 1.0);
-        let test_point = Point3::new(1.25, 1.25, 0.25);
-        let unoriented_edges: Vec<_> = cube.unoriented_edges_iter().collect();
-
-        // on the edge
-        let point_on_mesh_correct = Point3::new(1.0, 1.0, 0.25);
-        let pulled_point_on_mesh_calculated =
-            pull_point_to_mesh(&test_point, &cube, &unoriented_edges);
-
-        assert_eq!(pulled_point_on_mesh_calculated.point, point_on_mesh_correct);
-        assert_eq!(
-            pulled_point_on_mesh_calculated.distance,
-            nalgebra::distance(&test_point, &pulled_point_on_mesh_calculated.point),
-        );
-    }
-
-    #[test]
-    fn test_pull_point_to_mesh_cube_point_on_face() {
-        let cube = primitive::create_cube_sharp([0.0, 0.0, 0.0], 1.0);
-        let test_point = Point3::new(0.0, 0.0, 1.0);
-        let unoriented_edges: Vec<_> = cube.unoriented_edges_iter().collect();
-
-        let pulled_point_on_mesh_calculated =
-            pull_point_to_mesh(&test_point, &cube, &unoriented_edges);
-
-        assert_eq!(pulled_point_on_mesh_calculated.point, test_point);
-        assert_eq!(
-            pulled_point_on_mesh_calculated.distance,
-            nalgebra::distance(&test_point, &pulled_point_on_mesh_calculated.point),
-        );
-    }
-
-    #[test]
-    fn test_pull_point_to_mesh_cube_point_on_edge() {
-        let cube = primitive::create_cube_sharp([0.0, 0.0, 0.0], 1.0);
-        let test_point = Point3::new(1.0, 1.0, 0.0);
-        let unoriented_edges: Vec<_> = cube.unoriented_edges_iter().collect();
-
-        let pulled_point_on_mesh_calculated =
-            pull_point_to_mesh(&test_point, &cube, &unoriented_edges);
-
-        assert_eq!(pulled_point_on_mesh_calculated.point, test_point);
-        assert_eq!(
-            pulled_point_on_mesh_calculated.distance,
-            nalgebra::distance(&test_point, &pulled_point_on_mesh_calculated.point),
-        );
-    }
-
-    #[test]
-    fn test_is_point_on_plane_returns_true_for_point_on_plane() {
-        let plane_origin = Point3::new(1.0, 0.0, 0.0);
-        let plane_normal = Vector3::new(1.0, 0.0, 0.0);
-        let plane = Plane::from_origin_and_normal(&plane_origin, &plane_normal);
-        let test_point = Point3::new(1.0, 1.0, 1.0);
-
-        assert!(is_point_on_plane(&test_point, &plane));
-    }
-
-    #[test]
-    fn test_is_point_on_plane_returns_false_for_point_elsewhere() {
-        let plane_origin = Point3::new(1.0, 0.0, 0.0);
-        let plane_normal = Vector3::new(1.0, 0.0, 0.0);
-        let plane = Plane::from_origin_and_normal(&plane_origin, &plane_normal);
-        let test_point = Point3::new(2.0, 1.0, 1.0);
-
-        assert!(!is_point_on_plane(&test_point, &plane));
-    }
-
-    #[test]
     fn test_ray_intersects_plane_seed_2() {
         let normal = Vector3::new(0.505588, 0.843833, -0.179794);
         let ray_vector = Vector3::new(-0.708348, 0.05881, -0.703409);
@@ -1138,63 +1075,7 @@ mod tests {
     }
 
     #[test]
-    fn test_pull_point_to_plane_for_point_on_plane_and_horizontal_plane() {
-        let test_point: Point3<f32> = Point3::origin();
-        let test_plane =
-            Plane::from_origin_and_normal(&Point3::origin(), &Vector3::new(0.0, 0.0, 1.0));
-
-        let point_pulled_to_plane_calculated = pull_point_to_plane(&test_point, &test_plane);
-
-        assert_eq!(point_pulled_to_plane_calculated.point, test_point);
-        assert_eq!(point_pulled_to_plane_calculated.distance, 0.0);
-    }
-
-    #[test]
-    fn test_pull_point_to_plane_for_point_on_plane_and_vertical_plane() {
-        let test_point: Point3<f32> = Point3::origin();
-        let test_plane =
-            Plane::from_origin_and_normal(&Point3::origin(), &Vector3::new(1.0, 0.0, 0.0));
-
-        let point_pulled_to_plane_calculated = pull_point_to_plane(&test_point, &test_plane);
-
-        assert_eq!(point_pulled_to_plane_calculated.point, test_point);
-        assert_eq!(point_pulled_to_plane_calculated.distance, 0.0);
-    }
-
-    #[test]
-    fn test_pull_point_to_plane_for_point_above_plane_and_horizontal_plane() {
-        let test_point = Point3::new(0.0, 0.0, 1.0);
-        let test_plane =
-            Plane::from_origin_and_normal(&Point3::origin(), &Vector3::new(0.0, 0.0, 1.0));
-
-        let point_pulled_to_plane_calculated = pull_point_to_plane(&test_point, &test_plane);
-
-        assert_eq!(
-            point_pulled_to_plane_calculated.point,
-            Point3::new(0.0, 0.0, 0.0)
-        );
-        assert_eq!(point_pulled_to_plane_calculated.distance, 1.0);
-    }
-
-    #[test]
-    fn test_pull_point_to_plane_for_arbitrary_point_and_arbitrary_plane() {
-        let test_point = Point3::new(2.5, 4.2, 1.8);
-        let test_plane = Plane::from_origin_and_normal(&Point3::origin(), &test_point.coords);
-
-        let distance_correct = nalgebra::distance(&test_point, &Point3::origin());
-
-        let point_pulled_to_plane_calculated = pull_point_to_plane(&test_point, &test_plane);
-
-        assert!(point_pulled_to_plane_calculated.point.coords.relative_eq(
-            &Vector3::new(0.0, 0.0, 0.0),
-            0.001,
-            0.001,
-        ));
-        assert_eq!(point_pulled_to_plane_calculated.distance, distance_correct);
-    }
-
-    #[test]
-    fn test_pull_is_point_on_line_unclamped_returns_true_for_point_on_start() {
+    fn test_is_point_on_line_unclamped_returns_true_for_point_on_start() {
         let line_start = Point3::new(-1.0, -1.0, -1.0);
         let line_end = Point3::new(1.0, 1.0, 1.0);
         let test_point = Point3::new(-1.0, -1.0, -1.0);
@@ -1207,7 +1088,7 @@ mod tests {
     }
 
     #[test]
-    fn test_pull_is_point_on_line_unclamped_returns_true_for_point_on_end() {
+    fn test_is_point_on_line_unclamped_returns_true_for_point_on_end() {
         let line_start = Point3::new(-1.0, -1.0, -1.0);
         let line_end = Point3::new(1.0, 1.0, 1.0);
         let test_point = Point3::new(1.0, 1.0, 1.0);
@@ -1220,7 +1101,7 @@ mod tests {
     }
 
     #[test]
-    fn test_pull_is_point_on_line_unclamped_returns_true_for_point_in_the_middle() {
+    fn test_is_point_on_line_unclamped_returns_true_for_point_in_the_middle() {
         let line_start = Point3::new(-1.0, -1.0, -1.0);
         let line_end = Point3::new(1.0, 1.0, 1.0);
         let test_point = Point3::new(0.0, 0.0, 0.0);
@@ -1233,7 +1114,7 @@ mod tests {
     }
 
     #[test]
-    fn test_pull_is_point_on_line_unclamped_returns_true_for_point_before() {
+    fn test_is_point_on_line_unclamped_returns_true_for_point_before() {
         let line_start = Point3::new(-1.0, -1.0, -1.0);
         let line_end = Point3::new(1.0, 1.0, 1.0);
         let test_point = Point3::new(-2.0, -2.0, -2.0);
@@ -1246,7 +1127,7 @@ mod tests {
     }
 
     #[test]
-    fn test_pull_is_point_on_line_unclamped_returns_true_for_point_after() {
+    fn test_is_point_on_line_unclamped_returns_true_for_point_after() {
         let line_start = Point3::new(-1.0, -1.0, -1.0);
         let line_end = Point3::new(1.0, 1.0, 1.0);
         let test_point = Point3::new(2.0, 2.0, 2.0);
@@ -1259,7 +1140,7 @@ mod tests {
     }
 
     #[test]
-    fn test_pull_is_point_on_line_unclamped_returns_false_for_point_elsewhere() {
+    fn test_is_point_on_line_unclamped_returns_false_for_point_elsewhere() {
         let line_start = Point3::new(-1.0, -1.0, -1.0);
         let line_end = Point3::new(1.0, 1.0, 1.0);
         let test_point = Point3::new(3.0, 2.0, 1.0);
@@ -1272,7 +1153,7 @@ mod tests {
     }
 
     #[test]
-    fn test_pull_is_point_on_line_clamped_returns_true_for_point_on_start() {
+    fn test_is_point_on_line_clamped_returns_true_for_point_on_start() {
         let line_start = Point3::new(-1.0, -1.0, -1.0);
         let line_end = Point3::new(1.0, 1.0, 1.0);
         let test_point = Point3::new(-1.0, -1.0, -1.0);
@@ -1285,7 +1166,7 @@ mod tests {
     }
 
     #[test]
-    fn test_pull_is_point_on_line_clamped_returns_true_for_point_on_end() {
+    fn test_is_point_on_line_clamped_returns_true_for_point_on_end() {
         let line_start = Point3::new(-1.0, -1.0, -1.0);
         let line_end = Point3::new(1.0, 1.0, 1.0);
         let test_point = Point3::new(1.0, 1.0, 1.0);
@@ -1298,7 +1179,7 @@ mod tests {
     }
 
     #[test]
-    fn test_pull_is_point_on_line_clamped_returns_true_for_point_in_the_middle() {
+    fn test_is_point_on_line_clamped_returns_true_for_point_in_the_middle() {
         let line_start = Point3::new(-1.0, -1.0, -1.0);
         let line_end = Point3::new(1.0, 1.0, 1.0);
         let test_point = Point3::new(0.0, 0.0, 0.0);
@@ -1311,7 +1192,7 @@ mod tests {
     }
 
     #[test]
-    fn test_pull_is_point_on_line_clamped_returns_false_for_point_before() {
+    fn test_is_point_on_line_clamped_returns_false_for_point_before() {
         let line_start = Point3::new(-1.0, -1.0, -1.0);
         let line_end = Point3::new(1.0, 1.0, 1.0);
         let test_point = Point3::new(-2.0, -2.0, -2.0);
@@ -1324,7 +1205,7 @@ mod tests {
     }
 
     #[test]
-    fn test_pull_is_point_on_line_clamped_returns_false_for_point_after() {
+    fn test_is_point_on_line_clamped_returns_false_for_point_after() {
         let line_start = Point3::new(-1.0, -1.0, -1.0);
         let line_end = Point3::new(1.0, 1.0, 1.0);
         let test_point = Point3::new(2.0, 2.0, 2.0);
@@ -1337,7 +1218,7 @@ mod tests {
     }
 
     #[test]
-    fn test_pull_is_point_on_line_clamped_returns_false_for_point_elsewhere() {
+    fn test_is_point_on_line_clamped_returns_false_for_point_elsewhere() {
         let line_start = Point3::new(-1.0, -1.0, -1.0);
         let line_end = Point3::new(1.0, 1.0, 1.0);
         let test_point = Point3::new(3.0, 2.0, 1.0);
